refactor: 위치 객체 기본 생성자 접근제어자 닫기 및 수정
